# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minutor API",
        "description": "Sistema de apontamento de horas e despesas",
        "contact": {
            "email": "admin@minutor.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/v1/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Listar permissões",
                "description": "Lista todas as permissões do sistema",
                "operationId": "3ed6e7ff721a632320b8fe3c4a8c11ee",
                "parameters": [
                    {
                        "name": "grouped",
                        "in": "query",
                        "description": "Agrupar permissões por categoria",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de permissões",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "hasNext": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "guard_name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Criar permissão",
                "description": "Cria uma nova permissão no sistema",
                "operationId": "1535f9bb11e0c73c0f5f12d65fae99b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "projects.create"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Permissão criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/permissions/{id}": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Visualizar permissão",
                "description": "Exibe detalhes de uma permissão específica",
                "operationId": "42c08bfb1da681cdaf0a595146308f0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da permissão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Atualizar permissão",
                "description": "Atualiza uma permissão existente",
                "operationId": "dbe1d1689f32bf257bd397f01803fd44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da permissão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "projects.update"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permissão atualizada com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Excluir permissão",
                "description": "Remove uma permissão do sistema",
                "operationId": "177c482370c65e974f14d3e72e055666",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da permissão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permissão excluída com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/permissions/grouped": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Listar permissões agrupadas",
                "description": "Lista permissões agrupadas por categoria (admin, projects, hours, etc.)",
                "operationId": "0187ba2d38796f64ea17f54879e2c17f",
                "responses": {
                    "200": {
                        "description": "Permissões agrupadas por categoria"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação Bearer obtido através do endpoint de login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}